style: add explicit size_t to int casting from length functions

+++ print_char.c
@@ -6,7 +6,7 @@
 /*   By: frodrig2 <frodrig2@students.42porto.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/29 21:27:54 by frodrig2          #+#    #+#             */
-/*   Updated: 2026/05/29 21:37:04 by frodrig2         ###   ########.fr       */
+/*   Updated: 2026/05/31 03:03:03 by frodrig2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,5 +18,5 @@ int	print_char(va_list *ap_ptr)
 
 	c = va_arg(*ap_ptr, int);
 	ft_putchar_fd(c, STDOUT_FILENO);
-	return (sizeof(unsigned char));
+	return ((int)sizeof(unsigned char));
 }
